mass objects now use CalcLens, still need to fix tests
