+++ test_autogalaxy/plane/test_plane.py
@@ -843,7 +843,7 @@ class TestAbstractPlaneProfiles:
             assert (deflections.binned[1, 0] == 0.0).all()
             assert (deflections.binned[0] == 0.0).all()
 
-    class TestLensingObject:
+    class TestCalcLens:
         def test__correct_einstein_mass_caclulated_for_multiple_mass_profiles__means_all_innherited_methods_work(
             self,
         ):
